+++ jest.config.js
@@ -25,8 +25,8 @@ module.exports = {
   },
 
   // The glob patterns Jest uses to detect test files
-  testMatch: ["<rootDir>/test/unit/**/*.spec.js"],
-  // testMatch: ["<rootDir>/test/unit/**/Plugin.spec.js"],
+  testMatch: ["<rootDir>/test/e2e/**/*.spec.js"],
+  // testMatch: ["<rootDir>/test/e2e/**/legacy-behaviors-api.spec.js"],
 
   // The test environment that will be used for testing
   testEnvironment: "node",