+++ test/e2e/fixtures/start-files-watch.js
@@ -1,6 +1,0 @@
-const path = require("path");
-const { startServer } = require("../support/utils");
-
-startServer(path.resolve(__dirname, "files-watch"), {
-  watch: true,
-});